Revision 3: Add Guava memoization and consolidate tests
- Use Suppliers.memoize() for JAVA codec to cache instances and reduce GC pressure
- RedisStreamCodecIntegrationTest into RedisStreamCodecTest

Benefits:
- Memoization: Same instance reused, avoiding repeated object creation
- Thread-safe: Guava's memoize() handles concurrency
- Cleaner tests: Single test file with proper lifecycle management
